+++ Tests/Subscribe.MassTransit.OutboxEF/Subscribe.MassTransit.OutboxEF.Application/IntegrationEventHandlers/OrderCreatedEventHandler.cs
@@ -16,7 +16,7 @@ namespace Subscribe.MassTransit.OutboxEF.Application.IntegrationEventHandlers
     {
         private readonly ILogger<OrderCreatedEventHandler> _logger;
 
-        [IntentManaged(Mode.Ignore)]
+        [IntentManaged(Mode.Merge)]
         public OrderCreatedEventHandler(ILogger<OrderCreatedEventHandler> logger)
         {
             _logger = logger;