+++ Tests/Subscribe.MassTransit.OutboxMemory/Subscribe.MassTransit.OutboxMemory.Application/IntegrationEventHandlers/BasketDeletedEventHandler.cs
@@ -16,7 +16,7 @@ namespace Subscribe.MassTransit.OutboxMemory.Application.IntegrationEventHandler
     {
         private readonly ILogger<BasketDeletedEventHandler> _logger;
 
-        [IntentManaged(Mode.Ignore)]
+        [IntentManaged(Mode.Merge)]
         public BasketDeletedEventHandler(ILogger<BasketDeletedEventHandler> logger)
         {
             _logger = logger;