+++ Tests/Subscribe.MassTransit.OutboxMemory/Subscribe.MassTransit.OutboxMemory.Application/IntegrationEventHandlers/BasketItemCreatedEventHandler.cs
@@ -16,7 +16,7 @@ namespace Subscribe.MassTransit.OutboxMemory.Application.IntegrationEventHandler
     {
         private readonly ILogger<BasketItemCreatedEventHandler> _logger;
 
-        [IntentManaged(Mode.Ignore)]
+        [IntentManaged(Mode.Merge)]
         public BasketItemCreatedEventHandler(ILogger<BasketItemCreatedEventHandler> logger)
         {
             _logger = logger;