+++ Tests/Subscribe.MassTransit.OutboxMemory/Subscribe.MassTransit.OutboxMemory.Application/IntegrationEventHandlers/RoleDeletedEventHandler.cs
@@ -16,7 +16,7 @@ namespace Subscribe.MassTransit.OutboxMemory.Application.IntegrationEventHandler
     {
         private readonly ILogger<RoleDeletedEventHandler> _logger;
 
-        [IntentManaged(Mode.Ignore)]
+        [IntentManaged(Mode.Merge)]
         public RoleDeletedEventHandler(ILogger<RoleDeletedEventHandler> logger)
         {
             _logger = logger;